Add files via upload
A study was conducted in order to predict stock prices using Recurrent Neural Networks (RNNs). In order to do this, Sony Corporation's stock price data was chosen. Several experiments on RNN as well as experiments on improved recurrent networks like Long Short Term Memory (LSTM) and Gated Recurrent Unit (GRU) were done in order to understand the fundamental differences between the models, any benefits of using one over the rest. In addition, an additional experiment was done where the stock data of AMD Inc. was considered along with the stock data of Sony in order to predict whether AMD's stock trend had any impact on Sony's stocks, from which, it was hypothesised that AMD's stock's might have had at least some degree of impact on Sony's. In general, both LSTM and GRU performed better than RNN in terms of the mean squared error and loss. The mode from the additional experiment achieved a validation MSE of 0.000031 and a test MSE of 0.00029.